+++ Keke_PY/evaluation/tree_vs_linear__time_based/some_graphs.py
@@ -27,7 +27,7 @@ tree_setups: List[Tuple[bool, bool, int]] = [
 def get_synopsis(
         trees: bool, tracked: bool, job_nr: int,
         select_best_by_batch: int = 0, show_performance_on_batch: int = -1,
-) -> Tuple[List[Tuple[int, Dict[int, float]]], str]:
+) -> Tuple[List[Tuple[float, Dict[int, float]]], str]:
     file_name: str = file_name_template.replace("JOB", str(job_nr)).replace("TREES", str(int(trees)))
     print(f"reading in data for '{file_name}'")
     with open(file_name) as file:
@@ -37,8 +37,8 @@ def get_synopsis(
         representation = TrackedRepresentation(representation)
         representation.load_from_lines(log_lines)
     problem_data: KekeProblem = KekeProblem.from_log_lines(representation, log_lines, time_dependent_performance_function=True)
-    timeline: List[Tuple[int, Tuple[int, int]]] = get_best_instance_on_batch_timeline(problem_data, select_best_by_batch)
-    plot_data: List[Tuple[int, Dict[int, float]]] = get_all_performances_from_timeline(problem_data, timeline)
+    timeline: List[Tuple[float, Tuple[int, int]]] = get_best_instance_on_batch_timeline(problem_data, select_best_by_batch)
+    plot_data: List[Tuple[float, Dict[int, float]]] = get_all_performances_from_timeline(problem_data, timeline)
     individual: np.ndarray = problem_data.past_instances_by_gen_and_index[timeline[-1][1]]
     if tracked:
         return plot_data, representation.serialize_untracked(individual)
@@ -47,8 +47,8 @@ def get_synopsis(
 
 
 def stretch_from_points(
-        data: List[Tuple[int, Dict[int, float]]],
-        step_size: int = 10,
+        data: List[Tuple[float, Dict[int, float]]],
+        step_size: float = 10,
 ) -> List[Dict[int, float]]:
     last_data_point: Dict[int, float] = None
     res: List[Dict[int, float]] = []
